2-3: clarify byte/string for assets and helpers - py3 ready
The native string type is different for py2 and py3. The strings which
the helpers work on are ASCII. So, it is most natural to write out
native strings when we write out source files. This leads to (mostly)
unmodified source, avoids b and u prefixes and gives the same output
with py2 and py3.

Only "exception": base64 really encodes into (7bit) byte strings, so we
make this clear with a b prefix. This avoids an additional decode() when
we use them with PIL. (Alternative: avoid the prefix, use more decode())

With this commit assets2py.py and compile.py are py3 ready